refactoring: 게임 템플릿 분리
- 재사용성을 높이기 위해 게임 템플릿과 선택 로직을 분리함
